--- src/DbLocalizationProvider.Storage.PostgreSQL/GetResourceHandler.cs
@@ -1,40 +1,0 @@
-// Copyright (c) Valdis Iljuconoks. All rights reserved.
-// Licensed under Apache-2.0. See the LICENSE file in the project root for more information
-
-using DbLocalizationProvider.Abstractions;
-using DbLocalizationProvider.Queries;
-
-namespace DbLocalizationProvider.Storage.PostgreSql
-{
-    /// <summary>
-    /// Retrieves single resource
-    /// </summary>
-    public class GetResourceHandler : IQueryHandler<GetResource.Query, LocalizationResource>
-    {
-        private readonly ConfigurationContext _configurationContext;
-
-        /// <summary>
-        /// Creates new instance of the handler.
-        /// </summary>
-        /// <param name="configurationContext">Configuration settings.</param>
-        public GetResourceHandler(ConfigurationContext configurationContext)
-        {
-            _configurationContext = configurationContext;
-        }
-
-        /// <summary>
-        /// Place where query handling happens
-        /// </summary>
-        /// <param name="query">This is the query instance</param>
-        /// <returns>
-        /// You have to return something from the query execution. Of course you can return <c>null</c> as well if you
-        /// will.
-        /// </returns>
-        public LocalizationResource Execute(GetResource.Query query)
-        {
-            var repository = new ResourceRepository(_configurationContext);
-
-            return repository.GetByKey(query.ResourceKey);
-        }
-    }
-}
